+++ src/lib.rs
@@ -8,6 +8,7 @@ pub mod freeze;
 pub mod include;
 pub mod install;
 pub mod list_of_listings;
+pub mod listing_ref;
 pub mod manifest;
 pub mod number;
 pub mod verify;